+++ lib/hypa.ex
@@ -3,6 +3,6 @@ defmodule Hypa do
   `Hypa` is a set of utilities for building applications that use [htmx](https://htmx.org/).
   It provides the following modules:
 
-  - `Hypa.Plug` - A `Plug` module that adds data from HTMX headers to `conn.assigns`.
+  - `Hypa.Plug` - A `Plug` module that adds data from HTMX request headers to `conn.assigns`.
   """
 end